generate-password/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cloud-sql-python-connector==1.18.2
colorama==0.4.6
cryptography==45.0.4
dnspython==2.7.0
frozenlist==1.7.0
google-auth==2.40.3
idna==3.10
multidict==6.6.3
pillow==11.3.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyMySQL==1.1.1
qrcode==8.2
requests==2.32.4
rsa==4.9.1
urllib3==2.5.0
yarl==1.20.1
